requirements.txt
asgiref==3.8.0
Django==4.2.11
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
